+++ src/test/java/nextstep/subway/Db/DatabaseCleanup.java
@@ -1,4 +1,4 @@
-package nextstep.subway;
+package nextstep.subway.Db;
 
 import com.google.common.base.CaseFormat;
 import org.springframework.beans.factory.InitializingBean;
