+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -3,7 +3,7 @@ package nextstep.subway.station;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.DatabaseCleanup;
+import nextstep.subway.Db.DatabaseCleanup;
 import nextstep.subway.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
